requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
chardet==3.0.4
click==8.1.7
dnspython==2.6.1
deep-translator
email_validator==2.1.1
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
frozenlist==1.4.1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.5.1
httpcore==0.9.1
httptools==0.6.1
httpx>=0.23
hyperframe==5.2.0
idna==2.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
motor
multidict==6.0.5
orjson==3.10.3
packaging==24.0
pluggy==1.5.0
pydantic==2.7.1
pydantic-extra-types==2.7.0
pydantic-settings==2.2.1
pydantic_core==2.18.2
Pygments==2.18.0
python-dotenv
pymongo==4.7.2
pytest==8.2.1
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
rfc3986==1.5.0
rich==13.7.1
shellingham==1.5.4
sniffio==1.3.1
starlette==0.37.2
tomli==2.0.1
typer==0.12.3
typing_extensions==4.12.0
ujson==5.10.0
uvicorn
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
yarl==1.9.4
